skills/slide-extractor/requirements.txt
python-pptx
pytesseract
easyocr
Pillow
numpy
opencv-python-headless
